requirements.txt
Django>=5.0
djangorestframework>=3.14
gunicorn>=22.0
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
